Switch to OO approach
players are now objects inherited and overriden from a player interface
class

+++ Hnefatafl/Assets/GameLoad.cs
@@ -9,26 +9,26 @@ public class GameLoad : MonoBehaviour {
 	void Start ()
     {
         //Set Default Game Condition
-        Assets.ApplicationModel.gameType = (int)Assets.ApplicationModel.GameType.Player_V_Player;
+        Assets.ApplicationModel.gameType = Assets.ApplicationModel.GameType.Player_V_Player;
 	}
 
     void OnGUI()
     {
         if (GUI.Button(new Rect(10, 70, 250, 30), "Player Vs Player"))
         {
-            Assets.ApplicationModel.gameType = (int)Assets.ApplicationModel.GameType.Player_V_Player;
+            Assets.ApplicationModel.gameType = Assets.ApplicationModel.GameType.Player_V_Player;
             SceneManager.LoadScene("main");
         }
 
         if (GUI.Button(new Rect(10, 100, 250, 30), "AI Vs Player"))
         {
-            Assets.ApplicationModel.gameType = (int)Assets.ApplicationModel.GameType.AI_V_Player;
+            Assets.ApplicationModel.gameType = Assets.ApplicationModel.GameType.AI_V_Player;
             SceneManager.LoadScene("main");
         }
 
         if (GUI.Button(new Rect(10, 130, 250, 30), "AI Vs AI"))
         {
-            Assets.ApplicationModel.gameType = (int)Assets.ApplicationModel.GameType.AI_V_AI;
+            Assets.ApplicationModel.gameType = Assets.ApplicationModel.GameType.AI_V_AI;
             SceneManager.LoadScene("main");
         }
     }
